+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 2.4.0
+
+-   Added method for generating v2 Signed Multipart Upload Urls `CreateSignedUrlV2`
+
+# 2.3.0
+
+-   Added support for generating signed Custom Domain and PixelBin CDN urls
+
 # 2.2.0
 
 -   Added worker and Custom Domain url support to `UrlToObj` & `ObjToUrl`